kafka/requirements.txt
Flask==3.1.2
Werkzeug>=3.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
boto3==1.34.0
firebase-admin==6.4.0
psycopg2-binary==2.9.9
openai==1.12.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.2.0
easyocr==1.7.0
confluent-kafka==2.5.3
google-auth==2.35.0
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.21.0
google-cloud-aiplatform==1.73.0
google-generativeai==0.7.2
moviepy==1.0.3
